+++ src/ContaoManager/Plugin.php
@@ -43,9 +43,6 @@ class Plugin implements BundlePluginInterface, ConfigPluginInterface
 
     public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig)
     {
-        if (class_exists('HeimrichHannot\EncoreBundle\HeimrichHannotContaoEncoreBundle')) {
-            $loader->load('@HeimrichHannotFormhybridCompatibilityBundle/config/config_encore.yml');
-        }
         $loader->load('@HeimrichHannotFormhybridCompatibilityBundle/config/services.yml');
     }
 }